+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/Stage.component.js
@@ -1,13 +1,13 @@
 // @flow
 import React, { type ComponentType, useState, useCallback } from 'react';
 import cx from 'classnames';
-import i18n from '@dhis2/d2-i18n';
 import { withStyles } from '@material-ui/core';
-import { spacersNum, IconAdd16, Button } from '@dhis2/ui';
+import { spacersNum } from '@dhis2/ui';
 import { StageOverview } from './StageOverview';
 import type { Props } from './stage.types';
 import { Widget } from '../../../Widget';
 import { StageDetail } from './StageDetail/StageDetail.component';
+import { StageCreateNewButton } from './StageCreateNewButton';
 
 const styles = {
     overview: {
@@ -15,7 +15,7 @@ const styles = {
         marginRight: spacersNum.dp16,
         marginBottom: spacersNum.dp24,
     },
-    button: {
+    buttonContainer: {
         margin: `0 ${spacersNum.dp12}px ${spacersNum.dp8}px ${spacersNum.dp12}px`,
     },
     buttonRow: {
@@ -66,18 +66,16 @@ export const StagePlain = ({ stage, events, classes, className, onCreateNew, rul
                     hiddenProgramStage={preventAddingNewEvents}
                     {...passOnProps}
                 /> : (
-                    <Button
-                        small
-                        icon={<IconAdd16 />}
-                        className={classes.button}
-                        dataTest="create-new-button"
-                        onClick={() => onCreateNew(id)}
-                    >
-                        {i18n.t('New {{ eventName }} event', {
-                            eventName: name,
-                            interpolation: { escapeValue: false },
-                        })}
-                    </Button>
+                    <div className={classes.buttonContainer}>
+                        <StageCreateNewButton
+                            onCreateNew={() => onCreateNew(id)}
+                            stageWriteAccess={stage.dataAccess.write}
+                            eventCount={events.length}
+                            repeatable={repeatable}
+                            preventAddingEventActionInEffect={preventAddingNewEvents}
+                            eventName={name}
+                        />
+                    </div>
                 )}
             </Widget>
         </div>